code/1_prepare_data/3_analyze_new_brenda/requirements.txt
absl-py==0.9.0
alfpy==1.0.6
asn1crypto==1.3.0
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.3
biopython==1.76
bleach==1.4.2
-e git+git@github.com:EngqvistLab/brenparse.git@ec796588d8e718d10ceb6b466ccad1d5ab5c523d#egg=brenparse
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
-e git+git@github.com:mengqvist/colcol.git@3de5da8a9f77f7bb741f22c657e58a76394dcec9#egg=colcol
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
-e git+git@github.com:EngqvistLab/DeepET_reps.git@0d8797896e9ebc2bbc1055b6203e720a561136ed#egg=DeepET
defusedxml==0.6.0
entrypoints==0.3
gast==0.2.2
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
html5lib==1.0.1
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
llvmlite==0.31.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.5
notebook==6.0.3
numba==0.48.0
numpy==1.18.1
olefile==0.46
opt-einsum==3.1.0
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plotly==4.5.4
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.4
psutil==5.7.0
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.5.2
pygpu==0.7.6
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.7.1
QtPy==1.9.0
requests==2.23.0
retrying==1.3.3
scikit-learn==0.20.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
soupsieve==2.0.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
Theano==0.9.0
-e git+https://github.com/EngqvistLab/Tome.git@84a8c3c5b01e5da5f3a2db1e72cf4273768b1105#egg=tome
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
umap-learn==0.3.10
-e git+https://github.com/EngqvistLab/UniRep50@30ea0221431ddf98b88e37c20f7855768114664e#egg=UniRep50
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wrapt==1.11.2
-e git+https://github.com/mengqvist/wsvg@cac2e9492a1cd01b30a6287410c25ac68f846a26#egg=wsvg
zipp==2.2.0
